fixup! JAMES-3902 moving the generated keystore to server/container/guice/
