+++ server/protocols/jmap-draft/pom.xml
@@ -118,10 +118,6 @@
         </dependency>
         <dependency>
             <groupId>${james.groupId}</groupId>
-            <artifactId>james-server-guice-common</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>${james.groupId}</groupId>
             <artifactId>james-server-jmap</artifactId>
             <version>${project.version}</version>
         </dependency>
